# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadowTrace OSINT Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type DashboardSummaryThreatLevel =
  (typeof DashboardSummaryThreatLevel)[keyof typeof DashboardSummaryThreatLevel];

export const DashboardSummaryThreatLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface ModuleActivity {
  module: string;
  count: number;
  percentage: number;
}

export type HistoryEntryStatus =
  (typeof HistoryEntryStatus)[keyof typeof HistoryEntryStatus];

export const HistoryEntryStatus = {
  completed: "completed",
  failed: "failed",
  pending: "pending",
} as const;

export interface HistoryEntry {
  id: string;
  module: string;
  query: string;
  timestamp: string;
  riskScore: number;
  status: HistoryEntryStatus;
}

export interface WeeklyDataPoint {
  day: string;
  searches: number;
  risk: number;
}

export interface DashboardSummary {
  totalSearches: number;
  recentInvestigations: number;
  riskScore: number;
  activeModules: number;
  threatLevel: DashboardSummaryThreatLevel;
  activityByModule: ModuleActivity[];
  recentActivity: HistoryEntry[];
  weeklyActivity: WeeklyDataPoint[];
}

export interface UsernameSearchBody {
  username: string;
}

export type PlatformResultStatus =
  (typeof PlatformResultStatus)[keyof typeof PlatformResultStatus];

export const PlatformResultStatus = {
  found: "found",
  not_found: "not_found",
  error: "error",
} as const;

export interface ProfileData {
  displayName?: string;
  bio?: string;
  followers?: number;
  following?: number;
  posts?: number;
  verified?: boolean;
  joinDate?: string;
}

export interface PlatformResult {
  platform: string;
  url: string;
  status: PlatformResultStatus;
  profileData?: ProfileData;
}

export interface UsernameSearchResult {
  username: string;
  totalFound: number;
  riskScore: number;
  platforms: PlatformResult[];
  summary: string;
}

export interface EmailIntelligenceBody {
  email: string;
}

export type EmailIntelligenceResultReputation =
  (typeof EmailIntelligenceResultReputation)[keyof typeof EmailIntelligenceResultReputation];

export const EmailIntelligenceResultReputation = {
  clean: "clean",
  suspicious: "suspicious",
  malicious: "malicious",
  unknown: "unknown",
} as const;

export type BreachResultSeverity =
  (typeof BreachResultSeverity)[keyof typeof BreachResultSeverity];

export const BreachResultSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface BreachResult {
  name: string;
  date: string;
  dataClasses: string[];
  severity: BreachResultSeverity;
}

export interface EmailIntelligenceResult {
  email: string;
  isValid: boolean;
  domain: string;
  mxRecords: string[];
  breaches: BreachResult[];
  disposable: boolean;
  riskScore: number;
  reputation: EmailIntelligenceResultReputation;
}

export interface DomainLookupBody {
  domain: string;
}

export interface WhoisData {
  registrar: string;
  createdDate: string;
  expiryDate: string;
  updatedDate: string;
  registrant: string;
  country: string;
  nameservers: string[];
}

export interface DnsRecords {
  A: string[];
  MX: string[];
  NS: string[];
  TXT: string[];
  CNAME: string[];
}

export interface SslInfo {
  issuer: string;
  validFrom: string;
  validTo: string;
  valid: boolean;
}

export interface DomainLookupResult {
  domain: string;
  whois: WhoisData;
  dnsRecords: DnsRecords;
  subdomains: string[];
  technologies: string[];
  riskScore: number;
  sslInfo: SslInfo;
}

export interface IpTrackBody {
  ip: string;
}

export type IpTrackResultThreatLevel =
  (typeof IpTrackResultThreatLevel)[keyof typeof IpTrackResultThreatLevel];

export const IpTrackResultThreatLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface IpTrackResult {
  ip: string;
  city: string;
  region: string;
  country: string;
  countryCode: string;
  latitude: number;
  longitude: number;
  isp: string;
  org: string;
  asn: string;
  timezone: string;
  isProxy: boolean;
  isVpn: boolean;
  isTor: boolean;
  riskScore: number;
  threatLevel: IpTrackResultThreatLevel;
  openPorts: number[];
}

export interface PhoneLookupBody {
  phone: string;
}

export type PhoneLookupResultLineType =
  (typeof PhoneLookupResultLineType)[keyof typeof PhoneLookupResultLineType];

export const PhoneLookupResultLineType = {
  mobile: "mobile",
  landline: "landline",
  voip: "voip",
  unknown: "unknown",
} as const;

export interface PhoneLookupResult {
  phone: string;
  valid: boolean;
  countryCode: string;
  country: string;
  carrier: string;
  lineType: PhoneLookupResultLineType;
  region: string;
  spamScore: number;
  riskScore: number;
}

export interface MetadataExtractBody {
  url?: string;
  filename?: string;
  fileType: string;
}

export type MetadataExtractResultMetadata = { [key: string]: unknown };

export interface ExifData {
  make?: string;
  model?: string;
  software?: string;
  dateTime?: string;
  exposureTime?: string;
  fNumber?: string;
  iso?: number;
}

export interface GpsCoordinates {
  latitude: number;
  longitude: number;
  altitude: number;
}

export interface MetadataExtractResult {
  filename: string;
  fileType: string;
  fileSize: string;
  metadata: MetadataExtractResultMetadata;
  exifData?: ExifData;
  gpsCoordinates?: GpsCoordinates;
  riskScore: number;
  warnings: string[];
}

export interface GenerateReportBody {
  title: string;
  target: string;
  modules: string[];
  includeHistory?: boolean;
}

export interface ReportResult {
  id: string;
  title: string;
  target: string;
  createdAt: string;
  riskScore: number;
  summary: string;
  modules: string[];
  downloadUrl: string;
}

export interface ReportSummary {
  id: string;
  title: string;
  target: string;
  createdAt: string;
  riskScore: number;
  modules: string[];
}

export interface AiChatBody {
  message: string;
  context?: string;
}

export interface AiChatResult {
  response: string;
  suggestions: string[];
}

export type GetSearchHistoryParams = {
  limit?: number;
  module?: string;
};