# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShadowTrace OSINT Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ip/track:
    post:
      operationId: trackIp
      tags: [ip]
      summary: IP geolocation and intelligence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IpTrackBody"
      responses:
        "200":
          description: IP intelligence result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IpTrackResult"
...
components:
  schemas:
...
    IpTrackBody:
      type: object
      properties:
        ip:
          type: string
      required: [ip]
    IpTrackResult:
      type: object
      properties:
        ip:
          type: string
        city:
          type: string
        region:
          type: string
        country:
          type: string
        countryCode:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        isp:
          type: string
        org:
          type: string
        asn:
          type: string
        timezone:
          type: string
        isProxy:
          type: boolean
        isVpn:
          type: boolean
        isTor:
          type: boolean
        riskScore:
          type: number
        threatLevel:
          type: string
          enum: [low, medium, high, critical]
        openPorts:
          type: array
          items:
            type: integer
      required: [ip, city, region, country, countryCode, latitude, longitude, isp, org, asn, timezone, isProxy, isVpn, isTor, riskScore, threatLevel, openPorts]
...